backend/requirements.txt
# MedGuard AI + AdvocAI — Merged Dependencies
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart
python-dotenv

# LLM Clients
groq>=0.12.0
requests

# PDF Processing
pypdf>=4.0.0
PyMuPDF>=1.24.0
easyocr>=1.7.0
reportlab>=4.0.0

# AI / ML
sentence-transformers>=3.0.0
torch

# Database
psycopg2-binary>=2.9.0

# Auth
bcrypt>=4.0.0
PyJWT>=2.8.0

# Data Models
pydantic>=2.0.0
pydantic[email]

# CORS & Utils
aiofiles
